requirements.txt
Django>=4.0
djangorestframework
torch
transformers
anthropic
google-generativeai
python-dotenv
requests
